refactor(memory): split LanceDB plugin entry
