+++ config/max-lines-baseline.txt
@@ -175,7 +175,6 @@ extensions/memory-core/src/short-term-promotion.test.ts
 extensions/memory-core/src/tools.test.ts
 extensions/memory-core/src/tools.ts
 extensions/memory-lancedb/index.test.ts
-extensions/memory-lancedb/index.ts
 extensions/memory-wiki/src/chatgpt-import.ts
 extensions/memory-wiki/src/cli.ts
 extensions/memory-wiki/src/compile.ts
